requirements.txt
opencv-python
pyserial
numpy
Flask
ultralytics